testing/web-platform/mozilla/tests/xml/blob-content-type.html

<!doctype html>
<meta charset=utf-8>
<title></title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>

function parseBlob(blob) {
  return new Promise(resolve => {
    let xhr = new XMLHttpRequest();
    xhr.open("GET", URL.createObjectURL(blob));
    xhr.onload = () => {
      resolve(xhr.getResponseHeader("Content-Type"));
    }
    xhr.send();
  });
}

promise_test(async (t) => {
  // Create a new blob with a valid content type
  let blob = new Blob(["<x></x>"], { type: "application/xml" });

  // The blob content type should survive
  let blobContentType = await parseBlob(blob);
  assert_equals(blobContentType, "application/xml");
}, "blob object with content type is passed through");

promise_test(async (t) => {
  // Create a new blob with a invalid content type
  let blob = new Blob(["<x></x>"], { type: "\0" });

  // The blob content type should be overridden to text/xml
  let blobContentType = await parseBlob(blob);
  assert_equals(blobContentType, "text/xml");
}, "blob object with invalid content type is given text/xml");

promise_test(async (t) => {
  // Create a new blob without a content type
  let blob = new Blob(["<x></x>"]);

  // The blob content type should be overridden to text/xml
  let blobContentType = await parseBlob(blob);
  assert_equals(blobContentType, "text/xml");
}, "blob object with empty content type is given text/xml");

promise_test(async (t) => {
  let blob = new Blob(["<x></x>"], { type: "application/xml" });

  // Create a slice of a blob without a content type
  let slice = blob.slice(0, 6);

  // The slice content type should be overridden to text/xml
  let sliceContentType = await parseBlob(slice);
  assert_equals(sliceContentType, "text/xml");
}, "blob slice with empty content type is given text/xml");

</script>
